feat: finish volume work logic
