+++ src/ui/MoviePlayer/Range/Range.module.scss
@@ -1,31 +1,12 @@
 @use '@styles/functions' as fn;
 
-$buttonSize: fn.toRem(40px);
-$imageSize: fn.toRem(20px);
 $rangeColor: #ffffff;
-
-.volume {
-  display: flex;
-  align-items: center;
-}
-
-.button {
-  width: $buttonSize;
-  height: $buttonSize;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: transparent;
-}
-
-.image {
-  width: $imageSize;
-  height: $imageSize;
-}
+$rangeWidth: fn.toRem(48px);
+$rangeTrackHeight: fn.toRem(4px);
 
 .range {
   position: relative;
-  width: fn.toRem(48px);
+  width: $rangeWidth;
   height: fn.toRem(15px);
   background-color: transparent;
   cursor: pointer;
@@ -39,40 +20,44 @@ $rangeColor: #ffffff;
   position: absolute;
   top: 50%;
   left: 0;
-  width: fn.toRem(48px);
-  height: fn.toRem(4px);
+  width: $rangeWidth;
+  height: $rangeTrackHeight;
   background-color: rgba($color: $rangeColor, $alpha: 0.32);
   border-radius: 2px;
   transform: translateY(-50%);
+  z-index: -3;
 }
 
 .hoverRange {
   position: absolute;
   top: 50%;
   left: 0;
-  width: fn.toRem(48px);
-  height: fn.toRem(4px);
+  width: $rangeWidth;
+  height: $rangeTrackHeight;
   background-color: rgba($color: $rangeColor, $alpha: 0.4);
   border-radius: 2px;
   transform: translateY(-50%);
+  z-index: -2;
 }
 
 .selectedRange {
   position: absolute;
   top: 50%;
   left: 0;
-  width: fn.toRem(48px);
-  height: fn.toRem(4px);
+  width: $rangeWidth;
+  height: $rangeTrackHeight;
   background-color: $rangeColor;
   border-radius: 2px;
   transform: translateY(-50%);
+  z-index: -1;
 }
 
 .hoverRangeValue {
   position: absolute;
-  top: fn.toRem(-12px);
+  top: fn.toRem(-8px);
   font-weight: 600;
   font-size: fn.toRem(10px);
   color: $rangeColor;
   opacity: 0;
+  z-index: -1;
 }
