--- src/ui/MoviePlayer/Volume/index.ts
@@ -1,1 +1,0 @@
-export { Volume } from './Volume'
