+++ src/ui/MoviePlayer/VolumeButton/index.ts
@@ -1,0 +1,1 @@
+export { VolumeButton } from './VolumeButton'
